+++ packages/wallet/frontend/src/pages/transfer/request.tsx
@@ -27,7 +27,7 @@ import { Label } from '@/ui/forms/Label'
 import { FieldError } from '@/ui/forms/FieldError'
 import { useSnapshot } from 'valtio'
 import { balanceState } from '@/lib/balance'
-import { AssetOP } from '@/lib/api/asset'
+import { AssetOP } from '@wallet/shared'
 
 type SelectTimeUnitOption = Omit<SelectOption, 'value'> & {
   value: TimeUnit
