+++ packages/wallet/shared/src/types/incomingPayment.ts
@@ -1,0 +1,5 @@
+export interface PaymentDetailsResponse {
+  assetCode: string
+  description?: string
+  value: number
+}
