+++ packages/wallet/shared/src/types/index.ts
@@ -1,3 +1,5 @@
 export * from './account'
 export * from './asset'
 export * from './rate'
+export * from './qoute'
+export * from './incomingPayment'
