+++ packages/wallet/shared/src/types/qoute.ts
@@ -1,0 +1,12 @@
+import { AssetOP } from '@/types/asset'
+
+type AmountProps = AssetOP & {
+  value: string
+}
+
+export interface QuoteResponse {
+  id: string
+  receiveAmount: AmountProps
+  debitAmount: AmountProps
+  fee?: AmountProps
+}
